+++ Editor/AssemblyInfo.cs
@@ -3,25 +3,32 @@
 [assembly: InternalsVisibleTo("Unity.FilmInternalUtilities.EditorWorkbench")]
 [assembly: InternalsVisibleTo("Unity.FilmInternalUtilities.EditorTests")]
 
+[assembly: InternalsVisibleTo("Unity.AnimeToolbox")]
 [assembly: InternalsVisibleTo("Unity.AnimeToolbox.Editor")]
 [assembly: InternalsVisibleTo("Unity.AnimeToolbox.EditorTests")]
 
+[assembly: InternalsVisibleTo("Unity.GreasePencil")]
 [assembly: InternalsVisibleTo("Unity.GreasePencil.Editor")]
 [assembly: InternalsVisibleTo("Unity.GreasePencil.EditorTests")]
 
+[assembly: InternalsVisibleTo("Unity.MaterialSwitch")]
 [assembly: InternalsVisibleTo("Unity.MaterialSwitch.Editor")]
 [assembly: InternalsVisibleTo("Unity.MaterialSwitch.EditorTests")]
 
+[assembly: InternalsVisibleTo("Unity.MeshSync")]
 [assembly: InternalsVisibleTo("Unity.MeshSync.Editor")]
 [assembly: InternalsVisibleTo("Unity.MeshSync.EditorTests")]
 
+[assembly: InternalsVisibleTo("Unity.SelectionGroups")]
 [assembly: InternalsVisibleTo("Unity.SelectionGroups.Tests")]
 [assembly: InternalsVisibleTo("Unity.SelectionGroups.EditorTests")]
 
+[assembly: InternalsVisibleTo("Unity.StreamingImageSequence")]
 [assembly: InternalsVisibleTo("Unity.StreamingImageSequence.Editor")]
 [assembly: InternalsVisibleTo("Unity.StreamingImageSequence.EditorTests")]
 
 
+[assembly: InternalsVisibleTo("Unity.VisualCompositor")]
 [assembly: InternalsVisibleTo("Unity.VisualCompositor.Editor")]
 [assembly: InternalsVisibleTo("Unity.VisualCompositor.EditorTests")]