+++ Editor/Scripts/FilmInternalUtilitiesEditorConstants.cs
@@ -2,7 +2,7 @@
 
 namespace Unity.FilmInternalUtilities.Editor {
 
-internal static class FilmInternalUtilitiesConstants  {
+internal static class FilmInternalUtilitiesEditorConstants  {
     internal const string PACKAGE_NAME = "com.unity.film-internal-utilities";
     public static readonly string PACKAGE_PATH = Path.Combine("Packages", PACKAGE_NAME);