+++ Runtime/Scripts/Utilities/ObjectUtility.cs
@@ -1,6 +1,8 @@
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Assertions;
+using Object = UnityEngine.Object;
 #if UNITY_EDITOR
 using UnityEditor;
 #endif
@@ -48,19 +50,18 @@ internal static class ObjectUtility {
     }
     
 //----------------------------------------------------------------------------------------------------------------------       
-    
     /// <summary>
     /// Create a GameObject with a Component
     /// </summary>
     /// <param name="goName">The name of the GameObject</param>
     /// <typeparam name="T">The type of the Component</typeparam>
     /// <returns>The newly created GameObject</returns>
+    [Obsolete] 
     internal static T CreateGameObjectWithComponent<T>(string goName) where T: Component {
         GameObject go        = new GameObject(goName);
         T          component = go.AddComponent<T>();
         return component;        
     }
-    
 }
 
 } //end namespace