+++ Tests/Editor/Scripts/Utilities/EditorTestsUtility.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections;
+using UnityEditor;
+
+namespace Unity.FilmInternalUtilities.EditorTests {
+
+internal static class EditorTestsUtility {
+                
+    internal static IEnumerator WaitForFrames(int numFrames) {
+        for (int i = 0; i < numFrames; ++i) {
+            yield return null;
+            
+        }        
+        Undo.IncrementCurrentGroup();
+    }        
+}
+
+} //end namespace